Separated Registration::RegCodesLoader
