+++ src/lib/registration/storage.rb
@@ -23,10 +23,7 @@
 #
 
 require "singleton"
-require "tempfile"
-
-require "yast"
-require "registration/autoinstall_io"
+require "registration/registration_codes_loader"
 
 module Registration
   # a module holding data needed during (auto)installation
@@ -34,81 +31,10 @@ module Registration
     # storage for changed repositories
     class RegCodes < Struct.new(:reg_codes)
       include Singleton
-      include Yast::AutoinstallIoInclude
-      include Yast
-      include Yast::I18n
-      include Yast::Logger
+      include RegCodesLoader
 
       def initialize
-        Yast.import "XML"
-        # TODO: package it better
-        initialize_autoinstall_io(self)
-
         self.reg_codes = reg_codes_from_usb_stick || {}
-        log.info "Registration codes read for #{reg_codes.keys.inspect}"
-      end
-
-      private
-
-      REGCODES_NAME_HANDLERS = {
-        "regcodes.xml" => :reg_codes_from_xml,
-        "regcodes.txt" => :reg_codes_from_txt
-      }
-
-      # @return [Hash{String => String},nil]
-      def reg_codes_from_usb_stick
-        REGCODES_NAME_HANDLERS.each do |name, handler|
-          with_tempfile(name) do |path|
-            if GetURL("usb:///#{name}", path)
-              codes = send(handler, path)
-              return codes if codes
-            end
-          end
-        end
-        nil
-      end
-
-      # @param template for tempfile name
-      # @yieldparam actual file name
-      def with_tempfile(pattern, &block)
-        tempfile = Tempfile.new(pattern)
-        block.call(tempfile.path)
-      ensure
-        tempfile.unlink
-      end
-
-      # @param filename [String]
-      # @return [Hash{String => String},nil]
-      def reg_codes_from_xml(filename)
-        return nil unless File.readable?(filename)
-        xml_hash = XML.XMLToYCPFile(filename)
-        parse_xml(xml_hash)
-      end
-
-      # @param xml_hash [Hash] as used in AY and returned by XML.XMLToYCPFile
-      # @return [Hash{String => String},nil]
-      def parse_xml(xml_hash)
-        suse_register = xml_hash.fetch("suse_register", {})
-        addons = suse_register.fetch("addons", [])
-        pairs = addons.map do |a|
-          [a.fetch("name", ""), a.fetch("reg_code", "")]
-        end
-        Hash[pairs]
-      end
-
-      # The format is: lines with a key and a value,
-      #   separated by white space
-      # @param filename [String]
-      # @return [Hash{String => String},nil]
-      def reg_codes_from_txt(filename)
-        return nil unless File.readable?(filename)
-        text = File.read(filename)
-        # TODO: report parse errors in log
-        # TODO: allow for DOS line endings
-        pairs = text.each_line.map do |l|
-          l.chomp.split(/\s+/, 2)
-        end
-        Hash[pairs.reject(&:empty?)]
       end
     end
